requirements.txt
playwright~=1.41.0
psycopg2~=2.9.9
python-dotenv~=1.0.0
pytest~=7.4.4
pandas~=2.1.4
wandb~=0.16.2
opencv-python~=4.9.0.80
configobj~=5.0.8